core/st: add UI for project suggestions
A projection suggestion is an item in a suggestion menu like a normal suggestion but will insert a substitution when chosen for completion. With this patch, projections appear with a small preview in the suggestions menu that should make the choice more intuitive for users.

Besides, on the class-side protocol of st substitutions, #instanceSuggestion is extracted which answers a factory block for a new instance, making it easier for subclasses to provide their suggestions.

Suggestions are added to the substitutions memoize, point, language box, regex, and toggled code.