babylonian/core: dispatch SBStTestCase results into history
This change is visible when you run an SBStTestCase from a browser editor. In the past, the test result icon in the containg browser's message list would not have been updated (as compared to running the same test via the message list menu > "run test") - now this works. See also: CodeHolder>>#testRunSuite:.